Create gen server to intermediate channel and api
Create a GenServer to communicate with the client socket and the meetup
api.
With this solution we can handle socket reconnection and page
refreshes.
The server will start the data collection from the api when it is
created, and the client can join any time and receive the previous
collected members, and then it will receive the new members when they
arrive to server. This will also make sure we start only one data
collection by user access token and group slug.